# backend/api/swagger.yaml
openapi: "3.0.0"
info:
  title: API Music Room Suricates 
  description: This is a MusicRoom server with all the requests available to the users.
    You can authentify yourself to get access to the requests.
  version: 1.0.0
servers:
  - url: "http://62.34.5.191:45559/api"
  - url: "http://localhost:45559/api"
  - url: "http://localhost:46000/api"
...
paths:
...
  /users/{userId}/friendsInvitations:
    get:
      tags:
        - "user"
      summary: "Get your friends invitations"
      operationId: "getFriendsInvitations"
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
      responses:
        responses:
        200:
          description: |
            __{code:__ '0'__, friendsInvitations:__ *friendsInvitationsArray*__}__ - `Get an array of the friends invitations`
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Couldn't find the user`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
  /users/{userId}/playlistsInvitations:
    get:
      tags:
        - "user"
      summary: "Get your playlists invitations"
      operationId: "getPlaylistsInvitations"
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: |
            __{code:__ '0'__, playlistsInvitations:__ *playlistsInvitationArray*__}__ - `Get an array of the playlists invitations`
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Couldn't find the user`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
  /users/{userId}/eventsInvitations:
    get:
      tags:
        - "user"
      summary: "Get your events invitations"
      operationId: "getEventsInvitations"
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: |
            __{code:__ '0'__, eventsInvitations:__ *eventsInvitationsArray*__}__ - `Get an array of the events invitations`
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Couldn't find the user`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
...
  /users/{userId}/player/switchRandom:
    post:
      tags:
      - "player"
      summary: "Switch the random button of the player"
      operationId: "switchRandom"
      responses:
        200:
          description: |
            __{code:__ '0'__, message:__ *successMessage*__, random:__ 0__}__ - `Random player set to 'false'!`
         
            __{code:__ '1'__, message:__ *successMessage*__, random:__ 1__}__ - `Random player set to 'true'!`
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Couldn't find the user!`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
  /users/{userId}/player/refreshPlayer:
    post:
      tags:
      - "player"
      summary: "Refresh the tracksURI in the player"
      operationId: "refreshPlayer"
      responses:
        200:
          description: |
            __{code:__ '0'__, tracksId:__ *tracksIdArray*__, index:__ *arrayIndex*__}__ - `Reactualise the context of the playing list'!`
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `typeId isn't defined!`

            __{code:__ '1'__, message:__ *errorMessage*__}__ - `type must be 'none' or 'event' or 'playlist'!`

            __{code:__ '2'__, message:__ *errorMessage*__}__ - `userId isn't defined!`

            __{code:__ '3'__, message:__ *errorMessage*__}__ - `trackId isn't defined!`

            __{code:__ '4'__, message:__ *errorMessage*__}__ - `Couldn't find the event!`

            __{code:__ '5'__, message:__ *errorMessage*__}__ - `The track ID isn't in this event!`

            __{code:__ '6'__, message:__ *errorMessage*__}__ - `Couldn't find the playlist!`

            __{code:__ '7'__, message:__ *errorMessage*__}__ - `The track ID isn't in this playlist!`

            __{code:__ '8'__, message:__ *errorMessage*__}__ - `Couldn't find the user!`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                trackId:
                  type: string
                type:
                  type: string
                typeId:
                  type: string
              required:
                - trackId
  /users/{userId}/player/getNextTrack:
    post:
      tags:
      - "player"
      summary: "Get the next track to play"
      operationId: "getNextTrack"
      responses:
        200:
          description: |
            __{code:__ '0'__, trackId:__ *trackId*__}__ - `Get the next track to play in the playing list'!`
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Couldn't find the user!`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
  /users/{userId}/player/getPreviousTrack:
    post:
      tags:
      - "player"
      summary: "Get the previous track to play"
      operationId: "getPreviousTrack"
      responses:
        200:
          description: |
            __{code:__ '0'__, trackId:__ *trackId*__}__ - `Get the previous track to play in the playing list'!`
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Couldn't find the user!`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []